requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
pyasn1==0.6.1
pysmb==1.2.10
tqdm==4.67.1
Werkzeug==3.1.3
